+++ vistas/pesca/pescaInicio.php
@@ -1,5 +1,11 @@
 <?php
 include_once '../../clases/conexion.php';
+session_start();
+
+if (!isset($_SESSION['id'])){
+	header("location:/sectorpesquero/vistas/validarlogin.php");
+}
+
 
 $cursor = $pesca->find();
 ?>
@@ -30,6 +36,11 @@ $cursor = $pesca->find();
 
     <div class="contenidoFormulario">
         <div class="card mt-3 mb-3 bg-light">
+        <p> <?php
+               
+               echo  'Bienvenido '.  $_SESSION['usuario'];
+              
+                ?></p>
             <div class="card-header text-center" >
                 <h5>Pesca </h5>
             </div>
@@ -55,7 +66,13 @@ $cursor = $pesca->find();
                                     <td><?php echo $document['horarioSalida']; ?></td>
                                     <!-- Delete Button -->
                                     <td style="text-align:right">
+                                    <?php
+                                      if ($_SESSION['name']=="admin"){
+                                      ?>
                                         <a href="../../controladores/pesca/pescaEliminar.php?id=<?php echo $document['_id']; ?>" class="btn btn-danger"><i class="fa fa-trash-o" aria-hidden="true"></i></a>
+                                        <?php
+                                      }
+                                        ?>
                                         <a href="pescaActualizar.php?id=<?php echo $document['_id']; ?>" class="btn btn-warning"><i class="fa fa-pencil" aria-hidden="true"></i></a>
                                     </td>
                                 </tr>
