+++ vistas/tripulante/trupulanteInicio.php
@@ -1,5 +1,9 @@
 <?php
 include_once '../../clases/conexion.php';
+session_start();
+if (!isset($_SESSION['id'])){
+	header("location:/sectorpesquero/vistas/validarlogin.php");
+}
 
 $cursor = $tripulante->find();
 $id="";
@@ -32,6 +36,11 @@ $id="";
     <div class="contenidoFormulario">
 
         <div class="card mt-3 mb-3 bg-light">
+        <p> <?php
+               
+               echo  'Bienvenido '.  $_SESSION['usuario'];
+              
+                ?></p>
             <div class="card-header text-center">
                 <h5>Tripulación </h5>
             </div>
@@ -64,7 +73,13 @@ $id="";
                                     <td><?php echo $document['correo']; ?></td>
                                     <!-- Delete Button -->
                                     <td style="text-align:right">
+                                    <?php
+                                      if ($_SESSION['name']=="admin"){
+                                      ?>
                                         <a href="../../controladores/tripulante/tripulanteEliminar.php?id=<?php echo $document['_id']; ?>" class="btn btn-danger"><i class="fa fa-trash-o" aria-hidden="true"></i></a>
+                                        <?php
+                                      }
+                                        ?>
                                         <a href="tripulanteActualizar.php?id=<?php echo $document['_id']; ?>" class="btn btn-warning"><i class="fa fa-pencil" aria-hidden="true"></i></a>
                                     </td>
                                 </tr>
